scroll listener moved to Base, for each page load
